# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion Garage API",
        "description": "Documentation de l'API Gestion Garage",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/api/vehicules": {
            "get": {
                "tags": [
                    "Vehicules"
                ],
                "summary": "Lister les vehicules",
                "operationId": "92f70b047aec0a1f198bd4070eebf006",
                "responses": {
                    "200": {
                        "description": "Liste des vehicules recuperee avec succes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Vehicules"
                ],
                "summary": "Creer un vehicule",
                "operationId": "fe0bdc03b05e9a4026e0bc1f1f532ee5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "immatriculation",
                                    "marque",
                                    "modele"
                                ],
                                "properties": {
                                    "client_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "immatriculation": {
                                        "type": "string",
                                        "example": "RC-1234-AB"
                                    },
                                    "marque": {
                                        "type": "string",
                                        "example": "Toyota"
                                    },
                                    "modele": {
                                        "type": "string",
                                        "example": "Corolla"
                                    },
                                    "annee": {
                                        "type": "integer",
                                        "example": 2022
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vehicule cree avec succes"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/vehicules/{id}": {
            "get": {
                "tags": [
                    "Vehicules"
                ],
                "summary": "Afficher un vehicule",
                "operationId": "bdc297f5b091ddaf6eb63bf5f87e1dbf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant du vehicule",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicule recupere avec succes"
                    },
                    "404": {
                        "description": "Vehicule introuvable"
                    }
                }
            },
            "put": {
                "tags": [
                    "Vehicules"
                ],
                "summary": "Modifier un vehicule",
                "operationId": "71d9f63efc1782da6a4c7890e0b47c57",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant du vehicule",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "immatriculation",
                                    "marque",
                                    "modele"
                                ],
                                "properties": {
                                    "client_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "immatriculation": {
                                        "type": "string",
                                        "example": "RC-5678-CD"
                                    },
                                    "marque": {
                                        "type": "string",
                                        "example": "Hyundai"
                                    },
                                    "modele": {
                                        "type": "string",
                                        "example": "Tucson"
                                    },
                                    "annee": {
                                        "type": "integer",
                                        "example": 2023
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vehicule modifie avec succes"
                    },
                    "404": {
                        "description": "Vehicule introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Vehicules"
                ],
                "summary": "Supprimer un vehicule",
                "operationId": "7fd691db31af0a439b0df195730f011f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant du vehicule",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicule supprime avec succes"
                    },
                    "404": {
                        "description": "Vehicule introuvable"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Vehicules"
                ],
                "summary": "Modifier partiellement un vehicule",
                "operationId": "aacfd3dd8fda18aba44a4a0159f5ce9f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant du vehicule",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "client_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "immatriculation": {
                                        "type": "string",
                                        "example": "RC-5678-CD"
                                    },
                                    "marque": {
                                        "type": "string",
                                        "example": "Hyundai"
                                    },
                                    "modele": {
                                        "type": "string",
                                        "example": "Tucson"
                                    },
                                    "annee": {
                                        "type": "integer",
                                        "example": 2023
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vehicule modifie avec succes"
                    },
                    "404": {
                        "description": "Vehicule introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        }
    },
...
}